requirements.txt
asgiref==3.2.3
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto3==1.9.188
botocore==1.12.188
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-mapbox-location-field==1.4.0
django-secure==1.0.1
django-sslserver==0.22
djangorestframework==3.11.0
docutils==0.14
entrypoints==0.3
fake-useragent==0.1.11
Flask==1.1.1
future==0.17.1
google-api-core==1.16.0
google-auth==1.6.3
google-cloud-language==1.3.0
Google-Search-API==1.1.14
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.4.1
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.5.2
names-dataset==1.9.0
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.1
notebook==5.7.8
numpy==1.16.4
oauthlib==3.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pdfminer.six==20200124
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2==2.8.4
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycryptodome==3.9.6
Pygments==2.4.1
PyJWT==1.7.1
pylint==2.3.1
pyrsistent==0.15.2
python-dateutil==1.5
python-dotenv==0.11.0
python-social-auth==0.2.1
python3-openid==3.1.0
pytz==2019.3
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
selenium==2.53.6
Send2Trash==1.5.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sortedcontainers==2.1.0
soupsieve==1.9.1
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
tornado==6.0.2
traitlets==4.3.2
travis==0.0.3
typed-ast==1.4.0
Unidecode==1.1.0
urllib3==1.25.8
vcrpy==2.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==5.0.1
widgetsnbextension==3.4.2
wrapt==1.11.2
yarl==1.3.0